Create modular file structure for AudioTranscriptionClient
Split the monolithic AudioTranscriptionClient into separate files for better organization:

New Structure:
src/clients/audio_transcription/
├── __init__.py              # Package exports
├── models.py                # Pydantic models (TranscriptionRequest, TranscriptionResult)
├── protocols.py             # Protocol definitions (AudioFileValidator, etc.)
├── validators.py            # DefaultAudioFileValidator implementation
├── builders.py              # DefaultTranscriptionPayloadBuilder implementation
├── collectors.py            # DefaultTranscriptionResultCollector implementation
├── savers.py                # DefaultOutputSaver implementation
├── displays.py              # DefaultResultDisplay implementation
├── providers.py             # TranscriptionParameterProvider implementation
├── waiters.py               # QueueWaiter implementation
├── factories.py             # DefaultRunPodEndpointFactory implementation
└── client.py                # Main AudioTranscriptionClient class

Benefits:
- Single Responsibility: Each file has one clear purpose
- Better Organization: Easy to find specific functionality
- Improved Maintainability: Smaller, focused files
- Enhanced Testability: Can test individual components
- Better Navigation: Clear file structure
- Easier Extensions: Add new implementations in dedicated files

Backward Compatibility:
- Original audio_transcription_client.py now imports from new structure
- All existing imports continue to work
- No breaking changes to public API

All 66 unit tests pass
Maintains SOLID principles and clean code practices
