+++ src/clients/audio_transcription/__init__.py
@@ -1,0 +1,51 @@
+"""
+Audio transcription client package
+"""
+
+from .models import TranscriptionRequest, TranscriptionResult
+from .protocols import (
+    AudioFileValidator,
+    TranscriptionPayloadBuilder,
+    TranscriptionResultCollector,
+    OutputSaver,
+    ResultDisplay,
+    RunPodEndpoint,
+    RunPodEndpointFactory
+)
+from .validators import DefaultAudioFileValidator
+from .builders import DefaultTranscriptionPayloadBuilder
+from .collectors import DefaultTranscriptionResultCollector
+from .savers import DefaultOutputSaver
+from .displays import DefaultResultDisplay
+from .providers import TranscriptionParameterProvider
+from .waiters import QueueWaiter
+from .factories import DefaultRunPodEndpointFactory
+from .client import AudioTranscriptionClient
+
+__all__ = [
+    # Models
+    'TranscriptionRequest',
+    'TranscriptionResult',
+    
+    # Protocols
+    'AudioFileValidator',
+    'TranscriptionPayloadBuilder',
+    'TranscriptionResultCollector',
+    'OutputSaver',
+    'ResultDisplay',
+    'RunPodEndpoint',
+    'RunPodEndpointFactory',
+    
+    # Default Implementations
+    'DefaultAudioFileValidator',
+    'DefaultTranscriptionPayloadBuilder',
+    'DefaultTranscriptionResultCollector',
+    'DefaultOutputSaver',
+    'DefaultResultDisplay',
+    'TranscriptionParameterProvider',
+    'QueueWaiter',
+    'DefaultRunPodEndpointFactory',
+    
+    # Main Client
+    'AudioTranscriptionClient'
+]
