+++ src/clients/audio_transcription/providers.py
@@ -1,0 +1,27 @@
+"""
+Parameter provider implementations
+"""
+
+from typing import Optional, Dict
+from src.models import AppConfig
+
+
+class TranscriptionParameterProvider:
+    """Provider for transcription parameters with defaults"""
+    
+    def __init__(self, config: AppConfig):
+        self.config = config
+    
+    def get_parameters(self, model: Optional[str] = None, engine: Optional[str] = None) -> Dict[str, str]:
+        """Get transcription parameters with defaults"""
+        if self.config.transcription:
+            default_model = self.config.transcription.default_model
+            default_engine = self.config.transcription.default_engine
+        else:
+            default_model = "ivrit-ai/whisper-large-v3-ct2"
+            default_engine = "faster-whisper"
+        
+        return {
+            'model': model or default_model,
+            'engine': engine or default_engine
+        }
